requirements.txt
functions-framework==3.*
google-cloud-pubsub==2.13.11
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0